[Feat] Match info - Maps and vetoes
